requirements.txt
email-validator==1.1.3
Flask==2.0.1
flask-bootstrap==3.3.7.1
flask-login==0.5.0
Flask-SQLAlchemy==2.5.1
flask-wtf==0.15.1
python-dotenv==0.18.0
requests==2.26.0
twilio==6.63.0
